test: restructure screenshots to app module
