+++ app/src/screenshotTest/kotlin/com/divinelink/scenepeek/core/ui/ObfuscateSpoilersMenuItemScreenshots.kt
@@ -1,0 +1,11 @@
+package com.divinelink.scenepeek.core.ui
+
+import androidx.compose.runtime.Composable
+import com.divinelink.core.ui.Previews
+import com.divinelink.core.ui.components.dropdownmenu.ObfuscateTotalEpisodesMenuItemPreview
+
+@Previews
+@Composable
+fun ObfuscateSpoilersMenuItemScreenshots() {
+  ObfuscateTotalEpisodesMenuItemPreview()
+}
