+++ app/src/screenshotTest/kotlin/com/divinelink/scenepeek/core/ui/ScrollToTopButtonScreenshots.kt
@@ -1,0 +1,11 @@
+package com.divinelink.scenepeek.core.ui
+
+import androidx.compose.runtime.Composable
+import com.divinelink.core.ui.Previews
+import com.divinelink.core.ui.components.ScrollToTopButtonPreview
+
+@Previews
+@Composable
+fun ScrollToTopButtonScreenshots() {
+  ScrollToTopButtonPreview()
+}
