+++ app/src/screenshotTest/kotlin/com/divinelink/scenepeek/core/ui/rating/IMDbRatingItemScreenshots.kt
@@ -1,0 +1,11 @@
+package com.divinelink.scenepeek.core.ui.rating
+
+import androidx.compose.runtime.Composable
+import com.divinelink.core.ui.Previews
+import com.divinelink.core.ui.rating.IMDbRatingItemPreview
+
+@Previews
+@Composable
+fun IMDbRatingItemScreenshots() {
+  IMDbRatingItemPreview()
+}
