+++ app/src/screenshotTest/kotlin/com/divinelink/scenepeek/core/ui/rating/TMDBRatingItemScreenshots.kt
@@ -1,0 +1,11 @@
+package com.divinelink.scenepeek.core.ui.rating
+
+import androidx.compose.runtime.Composable
+import com.divinelink.core.ui.Previews
+import com.divinelink.core.ui.rating.TMDBRatingItemPreview
+
+@Previews
+@Composable
+fun TMDBRatingItemScreenshot() {
+  TMDBRatingItemPreview()
+}
