+++ app/src/screenshotTest/kotlin/com/divinelink/scenepeek/feature/details/person/ui/filter/PersonContentScreenshots.kt
@@ -1,0 +1,11 @@
+package com.divinelink.scenepeek.feature.details.person.ui.filter
+
+import androidx.compose.runtime.Composable
+import com.divinelink.core.ui.Previews
+import com.divinelink.feature.details.person.ui.filter.CreditsFilterBottomSheetContentPreview
+
+@Previews
+@Composable
+fun CreditsFilterBottomSheetContentScreenshots() {
+  CreditsFilterBottomSheetContentPreview()
+}
