+++ app/src/screenshotTest/kotlin/com/divinelink/scenepeek/feature/onboarding/SuccessTextScreenshots.kt
@@ -1,0 +1,11 @@
+package com.divinelink.scenepeek.feature.onboarding
+
+import androidx.compose.runtime.Composable
+import com.divinelink.core.ui.Previews
+import com.divinelink.feature.onboarding.ui.SuccessTextPreview
+
+@Previews
+@Composable
+fun SuccessTextScreenshots() {
+  SuccessTextPreview()
+}
