+++ app/src/screenshotTest/kotlin/com/divinelink/scenepeek/feature/settings/about/AboutCardScreenshots.kt
@@ -1,0 +1,11 @@
+package com.divinelink.scenepeek.feature.settings.about
+
+import androidx.compose.runtime.Composable
+import com.divinelink.core.ui.Previews
+import com.divinelink.feature.settings.app.about.AboutCardPreview
+
+@Previews
+@Composable
+fun AboutCardScreenshots() {
+  AboutCardPreview()
+}
