+++ app/src/screenshotTest/kotlin/com/divinelink/scenepeek/feature/watchlist/WatchlistContentScreenshots.kt
@@ -1,0 +1,11 @@
+package com.divinelink.scenepeek.feature.watchlist
+
+import androidx.compose.runtime.Composable
+import com.divinelink.core.ui.Previews
+import com.divinelink.feature.watchlist.WatchlistContentPreview
+
+@Composable
+@Previews
+fun WatchlistContentScreenshots() {
+  WatchlistContentPreview()
+}
