+++ core/ui/src/main/kotlin/com/divinelink/core/ui/Previews.kt
@@ -8,7 +8,6 @@ import androidx.compose.ui.tooling.preview.Preview
 @Preview(name = "light mode", uiMode = Configuration.UI_MODE_NIGHT_NO)
 @Preview(name = "dark mode", uiMode = Configuration.UI_MODE_NIGHT_YES)
 @Preview(name = "accessibility", device = "spec:width=360dp,height=640dp,dpi=480", fontScale = 2.0f)
-@Preview(name = "api 30", uiMode = Configuration.UI_MODE_NIGHT_NO, apiLevel = 30)
 annotation class Previews
 
 @Retention(AnnotationRetention.RUNTIME)
