--- core/ui/src/screenshotTest/kotlin/com/divinelink/core/ui/ObfuscateSpoilersMenuItemScreenshots.kt
@@ -1,10 +1,0 @@
-package com.divinelink.core.ui
-
-import androidx.compose.runtime.Composable
-import com.divinelink.core.ui.components.dropdownmenu.ObfuscateTotalEpisodesMenuItemPreview
-
-@Previews
-@Composable
-fun ObfuscateSpoilersMenuItemScreenshots() {
-  ObfuscateTotalEpisodesMenuItemPreview()
-}
