--- core/ui/src/screenshotTest/kotlin/com/divinelink/core/ui/rating/TraktRatingItemScreenshots.kt
@@ -1,10 +1,0 @@
-package com.divinelink.core.ui.rating
-
-import androidx.compose.runtime.Composable
-import com.divinelink.core.ui.Previews
-
-@Previews
-@Composable
-fun TraktRatingItemScreenshots() {
-  TraktRatingItemPreview()
-}
