--- feature/credits/src/screenshotTest/kotlin/com/divinelink/feature/credits/CreditsContentScreenshots.kt
@@ -1,16 +1,0 @@
-package com.divinelink.feature.credits
-
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.tooling.preview.PreviewParameter
-import com.divinelink.core.ui.Previews
-import com.divinelink.feature.credits.provider.CreditsUiStateParameterProvider
-import com.divinelink.feature.credits.ui.CreditsContentPreview
-import com.divinelink.feature.credits.ui.CreditsUiState
-
-@Previews
-@Composable
-fun CreditsContentScreenshots(
-  @PreviewParameter(CreditsUiStateParameterProvider::class) uiState: CreditsUiState,
-) {
-  CreditsContentPreview(uiState)
-}
