--- feature/credits/src/screenshotTest/kotlin/com/divinelink/feature/credits/PersonItemScreenshots.kt
@@ -1,11 +1,0 @@
-package com.divinelink.feature.credits
-
-import androidx.compose.runtime.Composable
-import com.divinelink.core.ui.Previews
-import com.divinelink.feature.credits.ui.PersonItemPreview
-
-@Previews
-@Composable
-fun PersonItemScreenshots() {
-  PersonItemPreview()
-}
