--- feature/details/src/screenshotTest/kotlin/com/divinelink/feature/details/media/ui/DetailsContentScreenshots.kt
@@ -1,13 +1,0 @@
-package com.divinelink.feature.details.media.ui
-
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.tooling.preview.PreviewParameter
-import com.divinelink.core.ui.Previews
-
-@Composable
-@Previews
-fun DetailsContentScreenshots(
-  @PreviewParameter(DetailsViewStateProvider::class) viewState: DetailsViewState,
-) {
-  DetailsContentPreview(viewState)
-}
