--- feature/details/src/screenshotTest/kotlin/com/divinelink/feature/details/person/ui/filter/PersonContentScreenshots.kt
@@ -1,10 +1,0 @@
-package com.divinelink.feature.details.person.ui.filter
-
-import androidx.compose.runtime.Composable
-import com.divinelink.core.ui.Previews
-
-@Previews
-@Composable
-fun CreditsFilterBottomSheetContentScreenshots() {
-  CreditsFilterBottomSheetContentPreview()
-}
