--- feature/onboarding/src/screenshotTest/kotlin/com/divinelink/feature/onboarding/SuccessTextScreenshots.kt
@@ -1,11 +1,0 @@
-package com.divinelink.feature.onboarding
-
-import androidx.compose.runtime.Composable
-import com.divinelink.core.ui.Previews
-import com.divinelink.feature.onboarding.ui.SuccessTextPreview
-
-@Previews
-@Composable
-fun SuccessTextScreenshots() {
-  SuccessTextPreview()
-}
