--- feature/settings/src/screenshotTest/kotlin/com/divinelink/feature/settings/app/about/AboutCardScreenshots.kt
@@ -1,10 +1,0 @@
-package com.divinelink.feature.settings.app.about
-
-import androidx.compose.runtime.Composable
-import com.divinelink.core.ui.Previews
-
-@Previews
-@Composable
-fun AboutCardScreenshots() {
-  AboutCardPreview()
-}
