--- feature/watchlist/src/screenshotTest/kotlin/com/divinelink/feature/watchlist/WatchlistContentScreenshots.kt
@@ -1,10 +1,0 @@
-package com.divinelink.feature.watchlist
-
-import androidx.compose.runtime.Composable
-import com.divinelink.core.ui.Previews
-
-@Composable
-@Previews
-fun WatchlistContentScreenshots() {
-  WatchlistContentPreview()
-}
